added key ParentNormalizedTrait
